Изменил название пакетов, после чего баг с Moxy исчерпал сам себя
