--- app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/ProfileMvpView.kt
@@ -1,10 +1,0 @@
-package ru.shcherbakovDmitry.ss.androidTraineeEducation
-
-import com.arellomobile.mvp.MvpView
-
-interface ProfileMvpView : MvpView {
-
-    fun fillProfileScreen(profile: UserProfile)
-
-    fun showLoginScreen()
-}
