--- app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/SearchFieldObserver.kt
@@ -1,5 +1,0 @@
-package ru.shcherbakovDmitry.ss.androidTraineeEducation
-
-interface SearchFieldObserver {
-    fun requestContent(key: String?)
-}
