+++ app/src/main/java/ru/shcherbakovDmitry/ss/androidTraineeEducation/mainscreen/dataclasses/CharityEvent.kt
@@ -1,4 +1,4 @@
-package ru.shcherbakovDmitry.ss.androidTraineeEducation
+package ru.shcherbakovDmitry.ss.androidTraineeEducation.mainscreen.dataclasses
 
 //TODO: Есть такое чувство, что может в будущем пригодиться фабрика или билдер
 class CharityEvent
