+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:id="@+id/container"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".MainActivity">
+    tools:context=".mainscreen.MainActivity">
 
     <android.support.v7.widget.Toolbar
         android:id="@+id/toolbar"
