+++ app/src/main/res/layout/fragment_events_tab.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".EventTabFragment">
+    tools:context=".mainscreen.search.charityeventtab.EventTabFragment">
 
     <android.support.constraint.ConstraintLayout
         android:id="@+id/layout_list_background"
