+++ app/src/main/res/layout/fragment_organisation_tab.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".OrganisationTabFragment">
+    tools:context=".mainscreen.search.organisationtab.OrganisationTabFragment">
 
     <TextView
         android:id="@+id/text_search_info"
