+++ app/src/main/res/layout/fragment_search_screen.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".SearchFragment"
+    tools:context=".mainscreen.search.SearchFragment"
     android:background="@color/searchFragmentBackground">
 
     <android.support.v4.view.ViewPager
